+++ twilio/rest/conversations/v1/conversation_with_participants.py
@@ -24,6 +24,7 @@ from twilio.base.version import Version
 class ConversationWithParticipantsInstance(InstanceResource):
 
     class State(object):
+        INITIALIZING = "initializing"
         INACTIVE = "inactive"
         ACTIVE = "active"
         CLOSED = "closed"