+++ twilio/rest/flex_api/v1/insights_assessments_comment.py
@@ -13,7 +13,7 @@ r"""
 """
 
 from typing import Any, Dict, List, Optional, Union, Iterator, AsyncIterator
-from twilio.base import deserialize, values
+from twilio.base import values
 
 from twilio.base.instance_resource import InstanceResource
 from twilio.base.list_resource import ListResource
@@ -43,14 +43,14 @@ class InsightsAssessmentsCommentInstance(InstanceResource):
         self.account_sid: Optional[str] = payload.get("account_sid")
         self.assessment_sid: Optional[str] = payload.get("assessment_sid")
         self.comment: Optional[Dict[str, object]] = payload.get("comment")
-        self.offset: Optional[float] = deserialize.decimal(payload.get("offset"))
+        self.offset: Optional[str] = payload.get("offset")
         self.report: Optional[bool] = payload.get("report")
-        self.weight: Optional[float] = deserialize.decimal(payload.get("weight"))
+        self.weight: Optional[str] = payload.get("weight")
         self.agent_id: Optional[str] = payload.get("agent_id")
         self.segment_id: Optional[str] = payload.get("segment_id")
         self.user_name: Optional[str] = payload.get("user_name")
         self.user_email: Optional[str] = payload.get("user_email")
-        self.timestamp: Optional[float] = deserialize.decimal(payload.get("timestamp"))
+        self.timestamp: Optional[str] = payload.get("timestamp")
         self.url: Optional[str] = payload.get("url")
 
     def __repr__(self) -> str: