+++ twilio/rest/verify/v2/service/entity/challenge/__init__.py
@@ -39,6 +39,7 @@ class ChallengeInstance(InstanceResource):
     class FactorTypes(object):
         PUSH = "push"
         TOTP = "totp"
+        PASSKEYS = "passkeys"
 
     class ListOrders(object):
         ASC = "asc"